Added checks if id exists in abstract service for delete and update

+++ src/main/java/com/nhlstenden/student/vigmo/services/AbstractVigmoService.java
@@ -34,7 +34,7 @@ public class AbstractVigmoService<Repository extends JpaRepository<Entity, Long>
             return mapper.mapObject(dbObject.get(), dtoType);
         }
         else{
-            throw new DataNotFoundException("ScreenService could not find " + id);
+            throw new DataNotFoundException(getClass().getName() + " could not find " + id);
         }
     }
 
@@ -47,13 +47,23 @@ public class AbstractVigmoService<Repository extends JpaRepository<Entity, Long>
 
     @Override
     public void update(DTO dto, long id) {
-        Entity newObject = mapper.mapObject(dto, entityType);
-        newObject.setId(id);
-        repo.save(newObject);
+        Optional<Entity> dbObject = repo.findById(id);
+        if(dbObject.isPresent()){
+            Entity newObject = mapper.mapObject(dto, entityType);
+            newObject.setId(id);
+            repo.save(newObject);
+        }else{
+            throw new DataNotFoundException(getClass().getName() + " could not find " + id);
+        }
     }
 
     @Override
     public void delete(long id) {
-        repo.deleteById(id);
+        Optional<Entity> dbObject = repo.findById(id);
+        if(dbObject.isPresent()){
+            repo.deleteById(id);
+        }else{
+            throw new DataNotFoundException("ScreenService could not find " + id);
+        }
     }
 }
